prelimary definition of int8_t and uint8_t

+++ test/numeric/integers.cpp
@@ -1,4 +1,3 @@
-#include "include/Helpers/Math.h"
 #define BOOST_TEST_MODULE IntegerTests
 #include <boost/test/included/unit_test.hpp>
 #include <boost/type_index.hpp>
@@ -12,6 +11,7 @@
 
 #include "include/Algos/Competition.h"
 #include "include/Algos/Integer.h"
+#include "include/Helpers/Math.h"
 
 namespace
 {
@@ -281,10 +281,10 @@ BOOST_AUTO_TEST_CASE(test_all_integegral_v)
 
   // test_and_benchmark_ints<int8_t>(0);
   // test_and_benchmark_ints<uint8_t>(0);
-  test_and_benchmark_ints<int16_t>(0);
-  test_and_benchmark_ints<uint16_t>(0);
-  // test_and_benchmark_ints<int32_t>(0);
-  // test_and_benchmark_ints<uint32_t>(0);
+  // test_and_benchmark_ints<int16_t>(0);
+  // test_and_benchmark_ints<uint16_t>(0);
+  test_and_benchmark_ints<int32_t>(0);
+  test_and_benchmark_ints<uint32_t>(0);
   // test_and_benchmark_ints<int64_t>(0);
   // test_and_benchmark_ints<uint64_t>(0);
 }
